requirements.txt
pycryptodome
configparser
rsa
colorama